+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.1.2] - 2019-07-229
+## [0.1.3] - 2019-08-18
+- Added related_rare_word(s) functions.
+- Fixed issue where first and second words of poem could be too similar
+
+## [0.1.2] - 2019-07-29
 - Fixed readme so example image will show on Pypi
 
 ## [0.1.1] - 2019-07-29
@@ -9,7 +13,7 @@ All notable changes to this project will be documented in this file.
 - Added example image to readme.
 - Added datamuse_api_max option to phonetically_related_words function.
 
-### Changed
+### [0.1.0]
 - Renamed frequently_intratextually_coappearing_word(s) functions to contextually_linked_word(s).
 - Defined public api with __all___.
 - Revised readme's installation instructions.